refac: moved read_charges to input.
