+++ source/databricks/calculation-engine/package/calculation/wholesale/__init__.py
@@ -14,4 +14,3 @@
 from .tariff_calculators import calculate_tariff_price_per_ga_co_es
 from .subscription_calculators import calculate_daily_subscription_price
 from .fee_calculators import calculate_fee_charge_price
-from .charges_reader import read_charges
