+++ source/databricks/calculation-engine/package/calculation_input/__init__.py
@@ -21,5 +21,7 @@ from .grid_loss_responsible import (
     get_grid_loss_responsible,
     _get_all_grid_loss_responsible,
 )
-from .metering_point_periods import get_metering_point_periods_df
+
+# from .metering_point_periods import get_metering_point_periods_df
 from .get_calculation_input import get_calculation_input
+from .charges_reader import read_charges
